fix
See https://getcomposer.org/doc/04-schema.md for details on the schema
The lock file is not up to date with the latest changes in composer.json, it is recommended that you run `composer update`.
The version field is present, it is recommended to leave it out if the package is published on Packagist.